non-type template arguments of type double not yet supported

+++ include/Math/Exp.hpp
@@ -422,17 +422,17 @@ constexpr auto sigmoid(double x) -> double { return 1.0 / (1.0 + exp(-x)); }
 constexpr auto softplus(double x) -> double { return log1p(exp(x)); }
 constexpr auto logit(double x) -> double { return log(x / (1.0 - x)); }
 
-template <double l = 8.0> constexpr auto smax(auto x, auto y) {
+template <int l = 8> constexpr auto smax(auto x, auto y) {
   auto d = x > y ? y - x : x - y;
   auto o = x > y ? decltype(d)(x) : decltype(d)(y);
   return o + softplus(l * d) / l;
 }
 
-template <double l = 8.0> constexpr auto smax(auto x, auto y, auto z) {
+template <int l = 8> constexpr auto smax(auto x, auto y, auto z) {
   double m = std::max(std::max(value(x), value(y)), value(z));
   return m + log(exp(l * (x - m)) + exp(l * (y - m)) + exp(l * (z - m))) / l;
 }
-template <double l = 8.0> constexpr auto smin(auto x, auto y) {
+template <int l = 8> constexpr auto smin(auto x, auto y) {
   return smax<-l>(x, y);
 }
